update django project

+++ project/product/models.py
@@ -5,7 +5,7 @@ class Product(models.Model):
     name = models.CharField(max_length=255)
     description = models.TextField()
     location = models.CharField(max_length=50)
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=10, decimal_places=2)
     color = models.CharField(max_length=50)
     
     def __str__(self):
